Added project "apiGame1.1"
Created version 1.1 of the API Card Game project. Currently, only shows the possibilities with the API and project. This version updates some of the code to be more standard and programmer friendly, and also updates some of the bugs that cam up, such as the following:

-Posting an error after consecutive new deck calls
-Posting an error after clearing an empty hand
-Posting an error after clearing an already shuffled deck/hand

Next version to be expected: 2.0